Refine tagged transaction info and reorganize Shiny tests
